Organize PNG codec
