+++ lib/image_util.rb
@@ -9,13 +9,8 @@ module ImageUtil
   autoload :Color, "image_util/color"
   autoload :Image, "image_util/image"
   autoload :Util, "image_util/util"
-  autoload :LibPNG, "image_util/libpng"
 
-  module Encoder
-    autoload :PNG, "image_util/encoder/png"
-  end
-
-  module Decoder
-    autoload :PNG, "image_util/decoder/png"
+  module Codec
+    autoload :Libpng, "image_util/codec/libpng"
   end
 end
